pyproject.toml
[project]
name = "pym3tools"
version = "0.1.0"
description = """Data utilities for processing Moon Mineralogy Mapper Data from
the NASA Planetary Data System."""
authors = [
    {name = "Z.M. Vig",email = "zvig@umd.edu"}
]
keywords = ["hyperspectral", "processing", "spectral cube", "image", "M3"]
license = {text = "LICENSE"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy (>=2.3.2,<3.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "h5py (>=3.14.0,<4.0.0)",
    "rasterio (>=1.4.3,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "scipy (>=1.16.1,<2.0.0)",
    "scipy-stubs (>=1.16.1.1,<2.0.0.0)",
    "types-pyyaml (>=6.0.12.20250822,<7.0.0.0)"
]

include = [
    "pym3tools/PDSretrieval/data/L0_INDEX_OP1.LBL",
    "pym3tools/PDSretrieval/data/L0_INDEX_OP1.TAB",
    "pym3tools/PDSretrieval/data/L0_INDEX_OP2.LBL",
    "pym3tools/PDSretrieval/data/L0_INDEX_OP2.TAB",
    "pym3tools/PDSretrieval/data/L1B_INDEX.LBL",
    "pym3tools/PDSretrieval/data/L1B_INDEX.TAB",
    "pym3tools/PDSretrieval/data/L2_INDEX.LBL",
    "pym3tools/PDSretrieval/data/L2_INDEX.TAB",
    "pym3tools/selenography/data/cgs_moon_2000.prj",
    "pym3tools/selenography/data/equidistant_cylindrical_moon.prj"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
